layout/base/tests/test_bug749186.html

<!DOCTYPE HTML>
<html>







  <head>
    <title>Test for Bug 749186 (Crashtest)</title>
    <script type="application/javascript" src="/MochiKit/packed.js"></script>
    <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
    <script>
     function endTest() {
       ok(true, 'test finished without crashing');
       SimpleTest.finish();
     }

     function removeBoldStyle() {
       document.getElementById('b').removeAttribute('style');
       SpecialPowers.pushPrefEnv({'set': [['font.size.inflation.emPerLine', 0]]},endTest);
     }

     function startTest() {
       SimpleTest.waitForExplicitFinish();
       SimpleTest.requestFlakyTimeout("untriaged");
       SpecialPowers.pushPrefEnv({'set': [['font.size.inflation.emPerLine', 8]]},removeBoldStyle);
     }

     startTest();
    </script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  </head>

  <body>
    <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=749186">Bug 749186</a>
    <iframe id="a" style="display: none;"></iframe>
    <div id="b" style="display: inline;"></div>
  </body>
</html>
